+++ src/shell/panels/control_center/common.h
@@ -14,32 +14,32 @@ namespace control_center {
 
 inline constexpr float kPreferredPanelWidth = 932.0f;
 inline constexpr float kPreferredPanelHeight =
-    static_cast<float>(Style::controlHeightLg * 15 + Style::spaceLg + Style::spaceSm);
+    Style::controlHeightLg * 15 + Style::spaceLg + Style::spaceSm;
 inline constexpr float kSidebarWidthRatio = 162.0f / 932.0f;
 inline constexpr float kSidebarWidth = kPreferredPanelWidth * kSidebarWidthRatio;
-inline constexpr float kSidebarButtonWidth = kSidebarWidth - static_cast<float>(Style::spaceMd * 2);
-inline constexpr float kContentMinWidth = static_cast<float>(Style::controlHeightLg * 14);
-inline constexpr float kMediaSliderWidth = static_cast<float>(Style::controlHeightLg * 9);
-inline constexpr float kValueLabelWidth = static_cast<float>(Style::controlHeightLg + Style::spaceLg);
-inline constexpr float kHeaderReserveHeight = static_cast<float>(Style::controlHeightSm);
-inline constexpr float kNotificationListRightPadding = static_cast<float>(Style::spaceXs);
+inline constexpr float kSidebarButtonWidth = kSidebarWidth - Style::spaceMd * 2;
+inline constexpr float kContentMinWidth = Style::controlHeightLg * 14;
+inline constexpr float kMediaSliderWidth = Style::controlHeightLg * 9;
+inline constexpr float kValueLabelWidth = Style::controlHeightLg + Style::spaceLg;
+inline constexpr float kHeaderReserveHeight = Style::controlHeightSm;
+inline constexpr float kNotificationListRightPadding = Style::spaceXs;
 inline constexpr float kMediaNowCardMinHeight =
-    static_cast<float>(Style::controlHeightLg * 11 + Style::spaceSm * 2);
+    Style::controlHeightLg * 11 + Style::spaceSm * 2;
 inline constexpr float kMediaAudioCardMinHeight =
-    static_cast<float>(Style::controlHeightLg * 3 + Style::spaceMd * 2);
-inline constexpr float kMediaPlayerSelectHeight = static_cast<float>(Style::controlHeight);
-inline constexpr float kMediaControlsHeight = static_cast<float>(Style::controlHeightLg + Style::spaceXs);
-inline constexpr float kMediaPlayPauseHeight = static_cast<float>(Style::controlHeightLg + Style::spaceSm);
-inline constexpr float kMediaArtworkMinHeight = static_cast<float>(Style::controlHeightLg * 4);
-inline constexpr float kCalendarCellSizeMin = static_cast<float>(Style::controlHeightSm + Style::spaceXs);
-inline constexpr float kCalendarCellSizeMax = static_cast<float>(Style::controlHeightLg + Style::spaceLg);
+    Style::controlHeightLg * 3 + Style::spaceMd * 2;
+inline constexpr float kMediaPlayerSelectHeight = Style::controlHeight;
+inline constexpr float kMediaControlsHeight = Style::controlHeightLg + Style::spaceXs;
+inline constexpr float kMediaPlayPauseHeight = Style::controlHeightLg + Style::spaceSm;
+inline constexpr float kMediaArtworkMinHeight = Style::controlHeightLg * 4;
+inline constexpr float kCalendarCellSizeMin = Style::controlHeightSm + Style::spaceXs;
+inline constexpr float kCalendarCellSizeMax = Style::controlHeightLg + Style::spaceLg;
 inline constexpr float kCalendarCellSize = kCalendarCellSizeMax;
 inline constexpr float kCalendarWeekdayWidth = kCalendarCellSize;
-inline constexpr float kCalendarGridGap = static_cast<float>(Style::spaceSm);
-inline constexpr float kCalendarCardPadding = static_cast<float>(Style::spaceMd);
-inline constexpr float kCalendarNavButtonSize = static_cast<float>(Style::controlHeight);
-inline constexpr float kCalendarWeekdayRowHeight = static_cast<float>(Style::controlHeight);
-inline constexpr float kCalendarHeaderHeight = static_cast<float>(Style::controlHeightLg + Style::spaceSm);
+inline constexpr float kCalendarGridGap = Style::spaceSm;
+inline constexpr float kCalendarCardPadding = Style::spaceMd;
+inline constexpr float kCalendarNavButtonSize = Style::controlHeight;
+inline constexpr float kCalendarWeekdayRowHeight = Style::controlHeight;
+inline constexpr float kCalendarHeaderHeight = Style::controlHeightLg + Style::spaceSm;
 inline constexpr float kCalendarCardMinWidth =
     kCalendarCellSize * 7.0f + kCalendarGridGap * 6.0f + kCalendarCardPadding * 2.0f;
 inline constexpr float kCalendarCardMinHeight =
