+++ src/shell/panels/control_center/network_tab.cpp
@@ -11,7 +11,7 @@ void ControlCenterPanel::buildNetworkTab() {
   auto tab = std::make_unique<Flex>();
   tab->setDirection(FlexDirection::Vertical);
   tab->setAlign(FlexAlign::Start);
-  tab->setGap(static_cast<float>(Style::spaceSm));
+  tab->setGap(Style::spaceSm);
   m_tabContainers[tabIndex(TabId::Network)] = tab.get();
 
   auto card = std::make_unique<Flex>();
